requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
pillow>=9.5.0
python-dotenv>=1.0.0
drf-nested-routers>=0.93.4
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
redis>=5.0.0
PyJWT>=2.8.0
